requirements.txt
pytest==3.1.3
requests==2.18.1
bs4==0.0.1
python-slugify==1.2.4
